+++ Changelog.md
@@ -30,3 +30,5 @@
 - fixed the using of multiple additional services
 - added "Fragile" and "Oversized" services support
 - added a option to choose pickup points list style
+- added Pakettikauppa block in order edit page
+- fixed additional service usage according to delivery method